Fix a couple of visual studio nightly build failures.
- Add explicit cast of bool to int to silence a test warning.
- Add explicit cast of size_t to int for same in dump_obu.

+++ test/intrabc_test.cc
@@ -146,7 +146,7 @@ TEST(IntrabcTest, DvValidation) {
   tile.mi_col_start = 24 * MAX_MIB_SIZE;
   tile.mi_col_end = tile.mi_col_start + kTileMaxMibWidth * MAX_MIB_SIZE;
   for (int i = 0; i < static_cast<int>(GTEST_ARRAY_SIZE_(kDvCases)); ++i) {
-    EXPECT_EQ(kDvCases[i].valid,
+    EXPECT_EQ(static_cast<int>(kDvCases[i].valid),
               av1_is_dv_valid(kDvCases[i].dv, &tile,
                               tile.mi_row_start + kDvCases[i].mi_row_offset,
                               tile.mi_col_start + kDvCases[i].mi_col_offset,
